move main methods to separate scripts
