+++ open-trip-planner/src/main/scala/com/azavea/busplan/routing/GenerateStudentToExistingStop.scala
@@ -6,19 +6,14 @@ import org.apache.commons.csv.CSVRecord;
 import java.io._
 import scala.collection.JavaConverters._
 
-object StudentToExistingStop {
+object GenerateStudentToExistingStopCosts {
+
   def main(args: Array[String]): Unit = {
     val stopToLocation = FileInput.readNodes(args(0))
     val studentToLocation = FileInput.readNodes(args(1))
     val studentToInfo = FileInput.readStudentInfo(args(1))
     val graph = RouteGraph.loadGraph(args(2))
     val walkRouter = new RouteGenerator(graph, graph, "WALK", false)
-
-    // val stopToLocation = StudentToStopMatrix.locationMap(args(0))
-    // val studentToLocation = StudentToStopMatrix.locationMap(args(1))
-    // val studentToInfo = CsvIo.readStudentInfo(args(1))
-    // val walkRouter = new WalkRouteCost(RouteGraph.loadGraph(args(2)))
-
     val outputCsv = FileOutput.initializeCsv(args(3), List("student_id", "stop_id", "count", "distance"))
     val eachCost = studentToInfo.map {
       case (key, value) => {
