+++ open-trip-planner/src/main/scala/com/azavea/busplan/routing/StudentToStopRouting.scala
@@ -6,24 +6,7 @@ import org.apache.commons.csv.CSVRecord;
 import java.io._
 import scala.collection.JavaConverters._
 
-object StudentStopRouting {
-
-  def main(args: Array[String]): Unit = {
-    val stopToLocation = FileInput.readNodes(args(0))
-    val studentToLocation = FileInput.readNodes(args(1))
-    val studentToPossibleStops = FileInput.readEligibleStops(args(2))
-    val graph = RouteGraph.loadGraph(args(3))
-    val walkRouter = new RouteGenerator(graph, graph, "WALK", false)
-
-    val results = routeAllStudents(studentToPossibleStops, stopToLocation,
-      studentToLocation, walkRouter)
-    Serialization.write(args(4), results)
-    val studentToInfo = FileInput.readStudentInfo(args(1))
-    createStudentToStopCSV("/home/azavea/files/da_customer/bus_routing/data-wrangling/student-to-stop/eligibility/eligibility-25.csv", results, 1320, studentToInfo)
-    createStudentToStopCSV("/home/azavea/files/da_customer/bus_routing/data-wrangling/cost-matrix/student-stop-eligibility-40.csv", results, 2112, studentToInfo)
-    createStudentToStopCSV("/home/azavea/files/da_customer/bus_routing/data-wrangling/cost-matrix/student-stop-eligibility-50.csv", results, 2640, studentToInfo)
-    createStudentToStopCSV("/home/azavea/files/da_customer/bus_routing/data-wrangling/cost-matrix/student-stop-eligibility-100.csv", results, 5280, studentToInfo)
-  }
+object StudentToStopRouting {
 
   def createStudentToStopCSV(
     path: String,
@@ -88,25 +71,4 @@ object StudentStopRouting {
       .toList
     Map(studentId -> costList)
   }
-
-  // def getWalkCost(oneStop: String,
-  //   studentLocation: Location,
-  //   stopToLocation: Map[String, Location],
-  //   walkRouter: WalkRouteCost): (String, Double) = {
-  //   val stopLocation = stopToLocation(oneStop)
-  //   val distance = walkRouter.calculateCost(studentLocation.coord,
-  //     stopLocation.coord, 1510560000).distance
-  //   (oneStop, distance)
-  // }
-
-  // def locationMap(filePath: String): Map[String, Location] = {
-  //   val nodes = CsvIo.readCsv(filePath)
-  //   sequenceToMap(nodes)
-  // }
-
-  // def sequenceToMap(nodes: Seq[Location]): Map[String, Location] = {
-  //   nodes
-  //     .map { node => Map(node.id -> node) }
-  //     .reduce { (map1, map2) => map1 ++ map2 }
-  // }
 }
